strontium::machine::instruction

Enum Instruction

pub enum Instruction {

    Halt,
    Load {
        value: RegisterValue,
        register: String,
    },
    Move {
        source: String,
        destination: String,
    },
    Copy {
        source: String,
        destination: String,
    },
    Push {
        value: RegisterValue,
        destination: String,
    },
    Append {
        value: Vec<RegisterValue>,
        destination: String,
    },
    Calculate {
        method: CalculationMethod,
        operand1: String,
        operand2: String,
        destination: String,
    },
    Compare {
        method: ComparisonMethod,
        operand1: String,
        operand2: String,
        destination: String,
    },
    Bitwise {
        method: BitwiseMethod,
    },
    Jump {
        destination: u32,
    },
    JumpC {
        destination: u32,
        conditional_address: String,
    },
    Interrupt {
        interrupt: Interrupt,
    },
    Call {
        address: usize,
    },
    Return,
    StoreLocal {
        name: String,
        register: String,
    },
    LoadLocal {
        name: String,
        register: String,
    },
    LoadType {
        source: String,
        destination: String,
    },
    LabelTarget {
        id: usize,
    },
    JumpToLabel {
        id: usize,
    },
    JumpCToLabel {
        id: usize,
        conditional_address: String,
    },
    CallShim {
        method_name: String,
    },

}

Represents a callable machine instruction

Variants

Halt

Stop all further execution

Load

Load a numeric value into a register

Fields

value: RegisterValue

register: String

Move

Move a value from one register address to another

Fields

source: String

destination: String

Copy

Copy a value from a register to memory or vice versa

Fields

source: String

destination: String

Push

Add a value to an array in a register

Fields

value: RegisterValue

The value to be pushed

destination: String

The name of the array register to append to

Append

Add a list of values to an array in a register

Fields

value: Vec<RegisterValue>

The values to append

destination: String

The name of the array register to append to

Calculate

Perform a calculation on two registers and write the result to a third

Fields

method: CalculationMethod

operand1: String

operand2: String

destination: String

Compare

Fields

method: ComparisonMethod

operand1: String

operand2: String

destination: String

Bitwise

Fields

method: BitwiseMethod

Jump

Set the program counter to the given value

Fields

destination: u32

JumpC

Set the program counter to a value if the given byte has the value of 1

Fields

destination: u32

conditional_address: String

Interrupt

Set off an interrupt, for example to print a character to standard output

Fields

interrupt: Interrupt

Call

Fields

address: usize

The address to jump to (the start of the method body).

Return

StoreLocal

Store a value from a register into the current stack frame’s local variables.

Fields

name: String

The name of the local variable.

register: String

The register to read the value from.

LoadLocal

Load a value from the current stack frame’s local variables into a register.

Fields

name: String

The name of the local variable.

register: String

The register to store the value in.

LoadType

Load the runtime type tag of a register’s value into another register as Int64.

Fields

source: String

destination: String

LabelTarget

Fields

id: usize

JumpToLabel

Fields

id: usize

JumpCToLabel

Fields

id: usize

conditional_address: String

CallShim

Call a named dispatch shim; resolved to Call { address } during linking.

Fields

method_name: String

Implementations

impl Instruction

pub fn get_opcode(&self) -> Opcode