strontium::machine

Struct Strontium

pub struct Strontium {
    pub registers: Registers,
    pub ip: usize,
    pub bytecode_parser: BytecodeParser,
    pub call_stack: Vec<StackFrame>,
    pub debug: bool,
    /* private fields */
}

Fields
registers: Registers

Holds general-purpose registers for storing different types of values

ip: usize

Points to the next index in the buffer of the bytecode register.

bytecode_parser: BytecodeParsercall_stack: Vec<StackFrame>debug: bool

Whether to print debug output during execution.

Implementations

impl Strontium

pub fn new(debug: bool) -> Self

Create a new instance of the virtual machine

pub fn reset(&mut self)

Reset the VM state for a new execution.

pub fn abort_execution(&mut self)

Abort the currently running program while keeping loaded methods and registers intact.

pub fn push_bytecode(&mut self, bytes: Vec<u8>)

Append machine code to the array in the bytecode register.

pub fn push_instruction(&mut self, instruction: Instruction)

pub fn parse_instruction(&mut self) -> Result<Instruction, StrontiumError>

pub fn execute(&mut self) -> Result<bool, StrontiumError>

Execute a single instruction.

pub fn execute_until_eof(&mut self) -> Result<bool, StrontiumError>

pub fn execute_until_eof_cancellable( &mut self, cancellation: &CancellationToken, ) -> Result<bool, StrontiumError>

pub fn execute_until_eof_interruptible<F>( &mut self, should_interrupt: F, ) -> Result<bool, StrontiumError>
where F: FnMut() -> bool,

pub fn consume_u64(&mut self) -> Result<u64, StrontiumError>

Consume an unsigned 64-bit integer from the bytecode register.

This performs a lookahead on the bytecode register to read the next eight bytes and converts the bytes into a 64-bit integer value. The byte encoding within Strontium bytecode is always Little Endian.

pub fn consume_u32(&mut self) -> Result<u32, StrontiumError>

pub fn consume_u16(&mut self) -> Result<u16, StrontiumError>

pub fn consume_u8(&mut self) -> Result<u8, StrontiumError>

pub fn consume_i64(&mut self) -> Result<i64, StrontiumError>

pub fn consume_i32(&mut self) -> Result<i32, StrontiumError>

pub fn consume_i16(&mut self) -> Result<i16, StrontiumError>

pub fn consume_i8(&mut self) -> Result<i8, StrontiumError>

pub fn consume_f64(&mut self) -> Result<f64, StrontiumError>

pub fn consume_f32(&mut self) -> Result<f32, StrontiumError>

pub fn consume_bool(&mut self) -> Result<bool, StrontiumError>

pub fn consume_byte(&mut self) -> Result<u8, StrontiumError>

pub fn consume_string(&mut self) -> Result<String, StrontiumError>