Module strontium::machine::instruction

Structs

CalculateExecutor

Perform a calculation two registers and write the result to a third.

CallExecutor

HaltExecutor

Stop all code execution immediately.

Interrupt

A signal indicating that an event needs immediate attention. This enumeration contains the interrupt types supported by the virtual machine.

InterruptExecutor

Attend to an event that needs immediate attention.

LoadExecutor

Load a dynamically typed value into a named register.

ReturnExecutor

Enums

BitwiseMethod

BytecodeError

CalculationMethod

Basic arithmetic operations which act on two registers and write the result to a third.

ComparisonMethod

Instruction

Represents a callable machine instruction

InterruptKind

Traits

Executor