Enum strontium::machine::instruction::Instruction

pub enum Instruction {

    HALT,
    LOAD {
        value: RegisterValue,
        register: String,
    },
    MOVE {
        source: String,
        destination: String,
    },
    COPY {
        source: String,
        destination: String,
    },
    PUSH {
        value: RegisterValue,
        destination: String,
    },
    APPEND {
        value: Vec<RegisterValue>,
        destination: String,
    },
    CALCULATE {
        method: CalculationMethod,
        operand1: String,
        operand2: String,
        destination: String,
    },
    COMPARE {
        method: ComparisonMethod,
        operand1: String,
        operand2: String,
        destination: String,
    },
    BITWISE {
        method: BitwiseMethod,
    },
    JUMP {
        destination: usize,
    },
    JUMPC {
        destination: usize,
        conditional_address: String,
    },
    INTERRUPT {
        interrupt: Interrupt,
    },
    CALL {},
    RETURN {},

}

Represents a callable machine instruction

Variants

HALT

Stop all further execution

LOAD

Fields

value: RegisterValue

register: String

Load a numeric value into a register

MOVE

Fields

source: String

destination: String

Move a value from one register address to another

COPY

Fields

source: String

destination: String

Copy a value from a register to memory or vice versa

PUSH

Fields

value: RegisterValue

The value to be pushed

destination: String

The name of the array register to append to

Add a value to an array in a register

APPEND

Fields

value: Vec<RegisterValue>

The values to append

destination: String

The name of the array register to append to

Add a list of values to an array in a register

CALCULATE

Fields

method: CalculationMethod

operand1: String

operand2: String

destination: String

Perform a calculation on two registers and write the result to a third

COMPARE

Fields

method: ComparisonMethod

operand1: String

operand2: String

destination: String

BITWISE

Fields

method: BitwiseMethod

JUMP

Fields

destination: usize

Set the program counter to the given value

JUMPC

Fields

destination: usize

conditional_address: String

Set the program counter to a value if the given byte has the value of 1

INTERRUPT

Fields

interrupt: Interrupt

Set off an interrupt, for example to print a character to standard output

CALL

Fields

RETURN

Fields